# Dokumenti/OpenAPISpec/niapi-openapi3.yaml
openapi: 3.0.1
info:
  version: 0.0.1
  title: niapi
  description: Specifikacija APIja koje implementira API servis narodnih inicijativa Srbije
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/elektronske-narodne-inicijative/niapi/0.0.1
  - url: https://einicijative.gov.rs/niapi/v1
paths:
...
  /salteri/inicijativa/{idInicijative}:
    get:
      description: Upit detalja za inicijativu
      security:
        - basicAuth: []
      parameters:
        - name: idInicijative
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalji o inicijativi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalterskiUpitInicijativeOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /salteri/potpisnik/{jmbg}/potpis:
    post:
      description: Potpisnik potpisuje narodnu inicijativu na salteru
      security:
        - basicAuth: []
      parameters:
        - name: jmbg
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/jmbg'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalterskiPotpisZahtev'
      responses:
        '200':
          description: Uspesno potpisana inicijativa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalterskiPotpisOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /salteri/potpisnik/{jmbg}/inicijativa/{idInicijative}:
    get:
      description: Provera detalja potpisa za inicijativu
      security:
        - basicAuth: []
      parameters:
        - name: jmbg
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/jmbg'
        - name: idInicijative
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalji o potpisanoj inicijativi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalterskiUpitPotpisaInicijativeOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /salteri/potpisnik/{jmbg}/potpisi:
    get:
      description: Provera detalja potpisa za inicijativu
      security:
        - basicAuth: []
      parameters:
        - name: jmbg
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/jmbg'
      responses:
        '200':
          description: Lista potpisa za zadatog potpisnika
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalterskiUpitListePotpisaOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'

  #  ---------------------------------------
  # Metodi namenjeni potpisnicima, javni API
  #  ---------------------------------------
  /potpisnik/profil:
    get:
      description: Upit detalja profila potpisnika
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Detalji profila potpisnika
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotpisnikProfilOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /potpisnik/potpis:
    post:
      description: Potpisnik potpisuje narodnu inicijativu online
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InicijativaZaPotpis'
      responses:
        '200':
          description: Detalji o potpisanoj inicijativi sa overom aplikacije (digitalnim potpisom)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotpisnikPotpisOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /potpisnik/inicijativa/{idInicijative}:
    get:
      description: Provera detalja potpisa za inicijativu
      security:
        - bearerAuth: []
      parameters:
        - name: idInicijative
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Detalji o potpisanoj inicijativi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotpisnikUpitPotpisaInicijativeOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'
  /potpisnik/potpisi:
    get:
      description: Provera detalja potpisa za inicijativu
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista potpisa za zadatog potpisnika
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotpisnikUpitListePotpisaOdgovor'
        '400':
          $ref: '#/components/responses/400Error'
        '401':
          $ref: '#/components/responses/401Error'
        '500':
          $ref: '#/components/responses/500Error'

  #  --------------------------------------------
  # Metodi namenjeni ovlašćenim licima, javni API
  #  --------------------------------------------
...
components:
  securitySchemes:

    basicAuth:
      type: http
      scheme: basic

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  #  -------------------------------------------------------
  #  Definicije struktura podataka
  #  -------------------------------------------------------
  schemas:
...
    jmbg:
      type: string
      pattern: '^[0-9]{13}$'
      minLength: 13
      maxLength: 13
...
    InicijativaZaPotpis:
      type: object
      required:
        - idInicijative
      properties:
        idInicijative:
          type: integer
    PotpisInicijative:
      type: object
      required:
        - idInicijative
        - nazivInicijative
        - idPotpisa
        - trnZavodjenjaPotpisa
      properties:
        idInicijative:
          type: integer
        nazivInicijative:
          type: string
        idPotpisa:
          type: string
          format: uuid
        trnZavodjenjaPotpisa:
          type: string
          format: date-time
...
    SalterskiUpitInicijativeOdgovor:
      type: object
      required:
        - idInicijative
        - nazivInicijative
        - tipInicijative
        - nivoVlasti
        - datumAktiviranja
      properties:
        idInicijative:
          type: integer
        nazivInicijative:
          type: string
        tipInicijative:
          type: string
        nivoVlasti:
          type: string
        datumAktiviranja:
          type: string
          format: date
    SalterskiPotpisZahtev:
      type: object
      required:
        - idInicijative
        - podaciSaSalteraPotpisa
        - imeKljucaZaOveru
        - overaAplikacije
      properties:
        idInicijative:
          type: integer
        podaciSaSalteraPotpisa:
          type: string
        imeKljucaZaOveru:
          type: string
        overaAplikacije:
          type: string
    SalterskiPotpisOdgovor:
      type: object
      required:
        - idInicijative
        - idPotpisa
        - trnZavodjenjaPotpisa
      properties:
        idInicijative:
          type: integer
        idPotpisa:
          type: string
          format: uuid
        trnZavodjenjaPotpisa:
          type: string
          format: date-time
    SalterskiUpitPotpisaInicijativeOdgovor:
      type: object
      properties:
        potpis:
            $ref: '#/components/schemas/PotpisInicijative'
    SalterskiUpitListePotpisaOdgovor:
      type: object
      properties:
        potpisi:
          type: array
          items:
            $ref: '#/components/schemas/PotpisInicijative'

    #
    #  Strukture za transakcije potpisnika
    #
    PotpisnikProfilOdgovor:
      type: object
      required:
        - idPola
        - nazivOpstine
        - godinaRodjenja
      properties:
        idPola:
          type: string
          enum: [М,Ж]
        nazivOpstine:
          type: string
        godinaRodjenja:
          type: integer
    PotpisnikPotpisOdgovor:
      type: object
      required:
        - idInicijative
        - idPotpisa
        - trnZavodjenjaPotpisa
        - imeKljucaZaOveru
        - overaAplikacije
      properties:
        idInicijative:
          type: integer
        idPotpisa:
          type: string
          format: uuid
        trnZavodjenjaPotpisa:
          type: string
          format: date-time
        imeKljucaZaOveru:
          type: string
        overaAplikacije:
          type: string
    PotpisnikUpitPotpisaInicijativeOdgovor:
      type: object
      properties:
        potpis:
            $ref: '#/components/schemas/PotpisInicijative'
    PotpisnikUpitListePotpisaOdgovor:
      type: object
      properties:
        potpisi:
          type: array
          items:
            $ref: '#/components/schemas/PotpisInicijative'

    #
    #  Strukture za transakcije ovlašćenih lica
    #
...
  responses:
    400Error:
      description: Aplikativne greske (pogresan sadrzaj)
      content:
        application/json:
          schema:
            type: object
            properties:
              resultType:
                type: string
                enum:
                  - PogresanZahtev
                  - PogresniPodaciUZahtevu
                  - NepoznatPotpisnik
                  - NepoznatInicijator
                  - NepoznatoOvlascenoLice
              resultCode:
                type: string
    401Error:
      description: Neispravna ili istekla sesija
      content:
        application/json:
          schema:
            type: object
            properties:
              resultType:
                type: string
                enum:
                  - NeispravnaSesija
                  - IsteklaSesija
              resultCode:
                type: string
    500Error:
      description: Neocekivane programske greske
      content:
        application/json:
          schema:
            type: object
            properties:
              resultType:
                type: string
                enum:
                  - ProblemUObradi
              resultCode:
                type: string