# Implementacija/Java/niapi/src/main/java/rs/gov/mduls/einicijative/niapi/interfaces/PotpisnikApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.0.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package rs.gov.mduls.einicijative.niapi.interfaces;

import org.springframework.security.core.Authentication;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.oauth2.core.user.OAuth2User;
import rs.gov.mduls.einicijative.niapi.model.InicijativaZaPotpis;
import rs.gov.mduls.einicijative.niapi.model.PotpisnikPotpisOdgovor;
import rs.gov.mduls.einicijative.niapi.model.PotpisnikProfilOdgovor;
import rs.gov.mduls.einicijative.niapi.model.PotpisnikUpitListePotpisaOdgovor;
import rs.gov.mduls.einicijative.niapi.model.PotpisnikUpitPotpisaInicijativeOdgovor;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;

import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-09-17T13:23:05.888263500+01:00[Europe/London]")
@Validated
public interface PotpisnikApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /potpisnik/inicijativa/{idInicijative}
     * Provera detalja potpisa za inicijativu
     *
     * @param idInicijative  (required)
     * @return Detalji o potpisanoj inicijativi (status code 200)
     *         or Aplikativne greske (pogresan sadrzaj) (status code 400)
     *         or Neispravna ili istekla sesija (status code 401)
     *         or Neocekivane programske greske (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/potpisnik/inicijativa/{idInicijative}",
        produces = { "application/json" }
    )
    default ResponseEntity<PotpisnikUpitPotpisaInicijativeOdgovor> potpisnikInicijativaIdInicijativeGet(
            @AuthenticationPrincipal OAuth2User principal,
            @PathVariable("idInicijative") Long idInicijative
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"potpis\" : { \"idInicijative\" : 0, \"trnZavodjenjaPotpisa\" : \"2000-01-23T04:56:07.000+00:00\", \"nazivInicijative\" : \"nazivInicijative\", \"idPotpisa\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /potpisnik/potpis
     * Potpisnik potpisuje narodnu inicijativu online
     *
     * @param inicijativaZaPotpis  (required)
     * @return Detalji o potpisanoj inicijativi sa overom aplikacije (digitalnim potpisom) (status code 200)
     *         or Aplikativne greske (pogresan sadrzaj) (status code 400)
     *         or Neispravna ili istekla sesija (status code 401)
     *         or Neocekivane programske greske (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/potpisnik/potpis",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<PotpisnikPotpisOdgovor> potpisnikPotpisPost(
            @AuthenticationPrincipal OAuth2User principal,
            @Valid @RequestBody InicijativaZaPotpis inicijativaZaPotpis
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"idInicijative\" : 0, \"overaAplikacije\" : \"overaAplikacije\", \"trnZavodjenjaPotpisa\" : \"2000-01-23T04:56:07.000+00:00\", \"imeKljucaZaOveru\" : \"imeKljucaZaOveru\", \"idPotpisa\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /potpisnik/potpisi
     * Provera detalja potpisa za inicijativu
     *
     * @return Lista potpisa za zadatog potpisnika (status code 200)
     *         or Aplikativne greske (pogresan sadrzaj) (status code 400)
     *         or Neispravna ili istekla sesija (status code 401)
     *         or Neocekivane programske greske (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/potpisnik/potpisi",
        produces = { "application/json" }
    )
    default ResponseEntity<PotpisnikUpitListePotpisaOdgovor> potpisnikPotpisiGet(
            @AuthenticationPrincipal OAuth2User principal
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"potpisi\" : [ { \"idInicijative\" : 0, \"trnZavodjenjaPotpisa\" : \"2000-01-23T04:56:07.000+00:00\", \"nazivInicijative\" : \"nazivInicijative\", \"idPotpisa\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, { \"idInicijative\" : 0, \"trnZavodjenjaPotpisa\" : \"2000-01-23T04:56:07.000+00:00\", \"nazivInicijative\" : \"nazivInicijative\", \"idPotpisa\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /potpisnik/profil
     * Upit detalja profila potpisnika
     *
     * @return Detalji profila potpisnika (status code 200)
     *         or Aplikativne greske (pogresan sadrzaj) (status code 400)
     *         or Neispravna ili istekla sesija (status code 401)
     *         or Neocekivane programske greske (status code 500)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/potpisnik/profil",
        produces = { "application/json" }
    )
    ResponseEntity<PotpisnikProfilOdgovor> potpisnikProfilGet(
            Authentication auth
    );

}